+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 0.2.9dev
-
+* [Fix] Do not check if there's internet connection when stats are disabled (#55)
 ## 0.2.8 (2023-03-27)
 
 * [Fix] Adds LICENSE information to `setup.py`